+++ actions_bootstrap.ps1
@@ -7,30 +7,30 @@ Get-PackageProvider -Name Nuget -ForceBootstrap | Out-Null
 Set-PSRepository -Name PSGallery -InstallationPolicy Trusted
 
 # List of PowerShell Modules required for the build
-$modulesToInstall = [System.Collections.ArrayList]::new()
+$modulesToInstall = New-Object System.Collections.Generic.List[object]
 # https://github.com/PowerShell/Plaster
-$null = $modulesToInstall.Add(([PSCustomObject]@{
-    ModuleName    = 'Plaster'
-    ModuleVersion = '1.1.4'
-}))
+[void]$modulesToInstall.Add(([PSCustomObject]@{
+            ModuleName    = 'Plaster'
+            ModuleVersion = '1.1.4'
+        }))
 # https://github.com/pester/Pester
-$null = $modulesToInstall.Add(([PSCustomObject]@{
+[void]$modulesToInstall.Add(([PSCustomObject]@{
             ModuleName    = 'Pester'
             ModuleVersion = '5.5.0'
         }))
 # https://github.com/nightroman/Invoke-Build
-$null = $modulesToInstall.Add(([PSCustomObject]@{
+[void]$modulesToInstall.Add(([PSCustomObject]@{
             ModuleName    = 'InvokeBuild'
             ModuleVersion = '5.10.5'
         }))
 # https://github.com/PowerShell/PSScriptAnalyzer
-$null = $modulesToInstall.Add(([PSCustomObject]@{
+[void]$modulesToInstall.Add(([PSCustomObject]@{
             ModuleName    = 'PSScriptAnalyzer'
             ModuleVersion = '1.21.0'
         }))
 # https://github.com/PowerShell/platyPS
 # older version used due to: https://github.com/PowerShell/platyPS/issues/457
-$null = $modulesToInstall.Add(([PSCustomObject]@{
+[void]$modulesToInstall.Add(([PSCustomObject]@{
             ModuleName    = 'platyPS'
             ModuleVersion = '0.12.0'
         }))